# internal/server-client/v1/pkg/types.gen.go
// Package pkg provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package pkg

import (
	"encoding/json"
	"errors"
	"time"

	"github.com/dndev-xx/go-ninja-chat/internal/types"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for ErrorCode.
const (
	ErrorCodeBadRequest           ErrorCode = 400
	ErrorCodeCreateChatError      ErrorCode = 1000
	ErrorCodeDeleteChatError      ErrorCode = 1002
	ErrorCodeForbidden            ErrorCode = 403
	ErrorCodeInternalServerError  ErrorCode = 500
	ErrorCodeMessageNotFoundError ErrorCode = 1003
	ErrorCodeNotFound             ErrorCode = 404
	ErrorCodeUnauthorized         ErrorCode = 401
	ErrorCodeUpdateChatError      ErrorCode = 1001
	N1004                         ErrorCode = 1004
)

// Defines values for MessageSentEventEventType.
const (
	MessageSentEventEventTypeMessageSentEvent MessageSentEventEventType = "messageSentEvent"
)

// BaseEvent defines model for BaseEvent.
type BaseEvent struct {
	EventId   *types.EventID   `json:"eventId,omitempty"`
	EventType string           `json:"eventType"`
	MessageId *types.MessageID `json:"messageId,omitempty"`
	RequestId *types.RequestID `json:"requestId,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code contains HTTP error codes and specific business logic error codes (the last must be >= 1000).
	Code    ErrorCode `json:"code"`
	Details *string   `json:"details,omitempty"`
	Message string    `json:"message"`
}

// ErrorCode contains HTTP error codes and specific business logic error codes (the last must be >= 1000).
type ErrorCode int

// Event defines model for Event.
type Event struct {
	union json.RawMessage
}

// GetHistoryRequest defines model for GetHistoryRequest.
type GetHistoryRequest struct {
	Cursor   *string `json:"cursor,omitempty"`
	PageSize *int    `json:"pageSize,omitempty"`
}

// GetHistoryResponse defines model for GetHistoryResponse.
type GetHistoryResponse struct {
	Data MessagesPage `json:"data"`
}

// Message defines model for Message.
type Message struct {
	AuthorId types.UserID `json:"authorId"`

	// Body The content of the message.
	Body string `json:"body"`

	// CreatedAt Timestamp when the message was created.
	CreatedAt time.Time       `json:"createdAt"`
	Id        types.MessageID `json:"id"`
}

// MessageHeader defines model for MessageHeader.
type MessageHeader struct {
	AuthorID  *types.UserID    `json:"authorID,omitempty"`
	CreatedAt *time.Time       `json:"createdAt,omitempty"`
	MessageID *types.MessageID `json:"messageID,omitempty"`
}

// MessageSentEvent defines model for MessageSentEvent.
type MessageSentEvent struct {
	AuthorId types.UserID `json:"authorId"`

	// Body The content of the message.
	Body string `json:"body"`

	// CreatedAt Timestamp when the message was created.
	CreatedAt time.Time                 `json:"createdAt"`
	EventId   *types.EventID            `json:"eventId,omitempty"`
	EventType MessageSentEventEventType `json:"eventType"`

	// IsService Indicates if the message is a service message.
	IsService bool             `json:"isService"`
	MessageId *types.MessageID `json:"messageId,omitempty"`
	RequestId *types.RequestID `json:"requestId,omitempty"`
}

// MessageSentEventEventType defines model for MessageSentEvent.EventType.
type MessageSentEventEventType string

// MessagesPage defines model for MessagesPage.
type MessagesPage struct {
	Messages []Message `json:"messages"`

	// TotalCount Total number of messages available.
	TotalCount int `json:"totalCount"`
}

// SendMessageRequest defines model for SendMessageRequest.
type SendMessageRequest struct {
	MessageBody string `json:"messageBody" validate:"required,min=1,max=4000"`
}

// SendMessageResponse defines model for SendMessageResponse.
type SendMessageResponse struct {
	Data  *MessageHeader `json:"data,omitempty"`
	Error *Error         `json:"error,omitempty"`
}

// XRequestIDHeader defines model for XRequestIDHeader.
type XRequestIDHeader = types.RequestID

// PostV1GetHistoryParams defines parameters for PostV1GetHistory.
type PostV1GetHistoryParams struct {
	XRequestID XRequestIDHeader `json:"X-Request-ID"`
}

// PostV1SendMessageParams defines parameters for PostV1SendMessage.
type PostV1SendMessageParams struct {
	XRequestID XRequestIDHeader `json:"X-Request-ID"`
}

// PostV1GetHistoryJSONRequestBody defines body for PostV1GetHistory for application/json ContentType.
type PostV1GetHistoryJSONRequestBody = GetHistoryRequest

// PostV1SendMessageJSONRequestBody defines body for PostV1SendMessage for application/json ContentType.
type PostV1SendMessageJSONRequestBody = SendMessageRequest

// AsMessageSentEvent returns the union data inside the Event as a MessageSentEvent
func (t Event) AsMessageSentEvent() (MessageSentEvent, error) {
	var body MessageSentEvent
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromMessageSentEvent overwrites any union data inside the Event as the provided MessageSentEvent
func (t *Event) FromMessageSentEvent(v MessageSentEvent) error {
	v.EventType = "messageSentEvent"
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeMessageSentEvent performs a merge with any union data inside the Event, using the provided MessageSentEvent
func (t *Event) MergeMessageSentEvent(v MessageSentEvent) error {
	v.EventType = "messageSentEvent"
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t Event) Discriminator() (string, error) {
	var discriminator struct {
		Discriminator string `json:"eventType"`
	}
	err := json.Unmarshal(t.union, &discriminator)
	return discriminator.Discriminator, err
}

func (t Event) ValueByDiscriminator() (interface{}, error) {
	discriminator, err := t.Discriminator()
	if err != nil {
		return nil, err
	}
	switch discriminator {
	case "messageSentEvent":
		return t.AsMessageSentEvent()
	default:
		return nil, errors.New("unknown discriminator value: " + discriminator)
	}
}

func (t Event) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *Event) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}