# internal/server-event/v1/pkg/types.gen.go
// Package pkg provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package pkg

import (
	"encoding/json"
	"errors"
	"time"

	"github.com/dndev-xx/go-ninja-chat/internal/types"
	"github.com/oapi-codegen/runtime"
)

// Defines values for MessageSentEventEventType.
const (
	MessageSentEventEventTypeMessageSentEvent MessageSentEventEventType = "messageSentEvent"
)

// BaseEvent defines model for BaseEvent.
type BaseEvent struct {
	EventId   *types.EventID   `json:"eventId,omitempty"`
	EventType string           `json:"eventType"`
	MessageId *types.MessageID `json:"messageId,omitempty"`
	RequestId *types.RequestID `json:"requestId,omitempty"`
}

// Event defines model for Event.
type Event struct {
	union json.RawMessage
}

// MessageSentEvent defines model for MessageSentEvent.
type MessageSentEvent struct {
	AuthorId types.UserID `json:"authorId"`

	// Body The content of the message.
	Body string `json:"body"`

	// CreatedAt Timestamp when the message was created.
	CreatedAt time.Time                 `json:"createdAt"`
	EventId   *types.EventID            `json:"eventId,omitempty"`
	EventType MessageSentEventEventType `json:"eventType"`

	// IsService Indicates if the message is a service message.
	IsService bool             `json:"isService"`
	MessageId *types.MessageID `json:"messageId,omitempty"`
	RequestId *types.RequestID `json:"requestId,omitempty"`
}

// MessageSentEventEventType defines model for MessageSentEvent.EventType.
type MessageSentEventEventType string

// AsMessageSentEvent returns the union data inside the Event as a MessageSentEvent
func (t Event) AsMessageSentEvent() (MessageSentEvent, error) {
	var body MessageSentEvent
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromMessageSentEvent overwrites any union data inside the Event as the provided MessageSentEvent
func (t *Event) FromMessageSentEvent(v MessageSentEvent) error {
	v.EventType = "messageSentEvent"
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeMessageSentEvent performs a merge with any union data inside the Event, using the provided MessageSentEvent
func (t *Event) MergeMessageSentEvent(v MessageSentEvent) error {
	v.EventType = "messageSentEvent"
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t Event) Discriminator() (string, error) {
	var discriminator struct {
		Discriminator string `json:"eventType"`
	}
	err := json.Unmarshal(t.union, &discriminator)
	return discriminator.Discriminator, err
}

func (t Event) ValueByDiscriminator() (interface{}, error) {
	discriminator, err := t.Discriminator()
	if err != nil {
		return nil, err
	}
	switch discriminator {
	case "messageSentEvent":
		return t.AsMessageSentEvent()
	default:
		return nil, errors.New("unknown discriminator value: " + discriminator)
	}
}

func (t Event) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *Event) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}